Add spec implementation support RSpec-like specs
In Ruby, we have developed a particular style of testing that works well for us. In addition to the describe/context/it style inherited from RSpec, as of lately, we have switched to a style that uses RSpec's let/subject to define the environment of collaborators, and use refinements of these collaborators in subsequent blocks to specify the behavior.

This represents an attempt at obtaining something similar in Kotlin, while also making use of type information for the collaborators (but not letting them get away). This turned out to be quite a challenge, as Kotlin provides much weaker meta-programming abilities, and something like Ruby's `module_eval` simply does not exist.

With this implementation, a test looks something like this

```
@TestFactory
fun test() = describe<Calculator> {
  val initialState = support { 0 }
  val calculator = subject { Calculator(initialState()) }

  describe("#state") {
    it("returns the initial state") {
      assertEqual(calculator().state(), 0)
    }

    context("when initial state is negative") {
      refine(initialState) { old -> old - 1 }

      it("still returns the initial state") {
        assertEqual(calculator().state(), -1)
      }
    }
  }
}
```

Compared to RSpec style, there are only a few downsides. The main one is probably that an collaborator can only refer to the ones declared before it, which effectively means that the subject will be the last one. Another undesireable change is that you need to put parenthesis after each use of a collaborator.

The former problem is caused by collaborators being defined inside a function scope (rather than a class scope), which technically means that they are local constants rather than class properties. An alternative would be to move the declarations to a class scope, but this becomes rather verbose if you want the ability to introduce additional collaborators on the inside. You would end up with something looking like

```
open class Collaborators {
  val calculator = subject { Calculator(initialState) }
  val initialState = support { 0 }
}
@TestFactory
fun test() = describe<Collaborators,Calculator> {
  describe("#state") {
    it("returns the initial state") {
      assertEqual(calculator.state(), 0)
    }

    open class Collaborators2: Collaborators() {
      override val initialState = support { super.initialState - 1 }
    }
    context<Collaborators2>("when initial state is negative") {
      it("still returns the initial state") {
        assertEqual(calculator.state(), -1)
      }
    }
  }
}
```

An advantage (from an implementation perspective) of this approach is that you refinement is handled automatically by the native override declaration, but the disadvantage is that that it adds additional useless verbosity (the class names), and makes it easy to use the wrong set of collaborators, and making it more complicated to move code around (as you would also have to rearrange names that you don't care about). Another advantage of this approach is that the issue with the additional parentheses goes away as well.

The second issue with the additional parentheses is caused by the fact that we want the (refinable) defintion to be evaluated exactly once per test instance, and given that we use local constants, this can only be achieved by returning (something that type-wise looks like) instance methods, so that we can implicitly capture the test instance from the method receiver. In principle, we could have returned (something that looks like) instance property references as well, but whereas you can write `calculator()` to implicitly capture `this` with a instance method reference, you cannot write `calculator` to implicitly capture `this` with an instance property reference. Instead you have to write `calculator.get()`.

In principle, we could get away without capturing `this` from within the test scope, by using a delegated property (combined with injecting dynamically constructed classes to have one per test instance and SecurityManager to determine the active instance class from the call-stack/classContext), but that would fail (at runtime) if you e.g. spawned a thread in an example, and used the delegated property from that thread.