Add support for destructors for Spec collaborators
You can pass a destructor as the first argument to support and subject, which is a bit strange, but since we want to use the last argument for the constructor, there is not really any better way to put it. That being said, I suppose the typical use case would be to define a wrapper function that constructs the destructor, so it should be relatively problem free.

This also adds automatic destruction of autoclosables.