Disallow nesting spec operators inside it blocks
After performing various copy-paste operations I've managed to end up with it blocks nested in other it blocks, which previously compiled fine, and only generated a mysterious concurrent-modification error deep inside JUnit, with very limited context information.

This introduces copies of all the spec infrastructure methods inside the TestInstance with identical signatures to ensure that the original methods are properly shadowed. Initially I intended to implement each method to throw a run-time exception, but figured out that you could add a Deprecation annotation to turn it into a compile-time error instead of a runtime one.

Given that these result in compile-time errors on mistakes, it would be somewhat challenging to add tests for these.