Add integration-style test for akts script
This launches an inline akts script in a JVM scripting host.

However, given that the AktsScript base class is defined in the compilation output, we need to inject the classpath containing the base class into the compilation. There are more high-level classpath manipulations that can be performed, but they seemed to match only the resources compilation output directory, which doesn't really help, so uses a not-so-nice low level hack.